requirements.txt
Django~=3.2.7
djangorestframework~=3.12.4
psycopg2~=2.9.1
aiogram~=2.15
requests~=2.26.0
google-api-python-client~=2.23.0
google-auth-httplib2~=0.1.0
google-auth-oauthlib~=0.4.6